Changing byte writing method, convert raw dmg to %

+++ DD2-SbR-Mod/Models/Car.cs
@@ -79,7 +79,7 @@ namespace Sbr.Models
                     }
                 if(LapNumber == modConfig.lapLimit)
                     {
-                        rw.SetByte(RaceMemoryAddress, (byte)modConfig.map.LapsNumber);
+                        rw.SetByte(RaceMemoryAddress, modConfig.map.LapsNumber);
                     }
             }
             else
